+++ src/ErrorBoundary/ErrorBoundary.tsx
@@ -116,7 +116,7 @@ class ErrorBoundary extends React.Component<
                             problem we recommend restarting the application.
                         </p>
                         <Button
-                            large
+                            size="lg"
                             variant="primary"
                             onClick={() => getCurrentWindow().reload()}
                         >
@@ -167,7 +167,7 @@ class ErrorBoundary extends React.Component<
                         experienced it multiple times
                     </p>
                     <Button
-                        large
+                        size="lg"
                         variant="primary"
                         onClick={() =>
                             openUrl(